Renamed the library from ``PyTimeCode`` to ``timecode``
