+++ tests/test_timecode.py
@@ -1,11 +1,11 @@
 #!-*- coding: utf-8 -*-
 
 import unittest
-from pytimecode import PyTimeCode
+from timecode import Timecode
 
 
-class PyTimeCodeTester(unittest.TestCase):
-    """tests PyTimeCode class
+class TimecodeTester(unittest.TestCase):
+    """tests Timecode class
     """
 
     @classmethod
@@ -23,180 +23,180 @@ class PyTimeCodeTester(unittest.TestCase):
     def test_instance_creation(self):
         """test instance creation, none of these should raise any error
         """
-        PyTimeCode('23.98', '00:00:00:00')
-        PyTimeCode('24', '00:00:00:00')
-        PyTimeCode('25', '00:00:00:00')
-        PyTimeCode('29.97', '00:00:00:00')
-        PyTimeCode('30', '00:00:00:00')
-        PyTimeCode('50', '00:00:00:00')
-        PyTimeCode('59.94', '00:00:00:00')
-        PyTimeCode('60', '00:00:00:00')
-        PyTimeCode('ms', '03:36:09:230')
-        PyTimeCode('24', start_timecode=None, frames=12000)
+        Timecode('23.98', '00:00:00:00')
+        Timecode('24', '00:00:00:00')
+        Timecode('25', '00:00:00:00')
+        Timecode('29.97', '00:00:00:00')
+        Timecode('30', '00:00:00:00')
+        Timecode('50', '00:00:00:00')
+        Timecode('59.94', '00:00:00:00')
+        Timecode('60', '00:00:00:00')
+        Timecode('ms', '03:36:09:230')
+        Timecode('24', start_timecode=None, frames=12000)
 
-        PyTimeCode('23.98')
-        PyTimeCode('24')
-        PyTimeCode('25')
-        PyTimeCode('29.97')
-        PyTimeCode('30')
-        PyTimeCode('50')
-        PyTimeCode('59.94')
-        PyTimeCode('60')
-        PyTimeCode('ms')
-        PyTimeCode('24', frames=12000)
+        Timecode('23.98')
+        Timecode('24')
+        Timecode('25')
+        Timecode('29.97')
+        Timecode('30')
+        Timecode('50')
+        Timecode('59.94')
+        Timecode('60')
+        Timecode('ms')
+        Timecode('24', frames=12000)
 
     def test_repr_overload(self):
-        timeobj = PyTimeCode('24', '01:00:00:00')
+        timeobj = Timecode('24', '01:00:00:00')
         self.assertEqual('01:00:00:00', timeobj.__repr__())
 
-        timeobj = PyTimeCode('23.98', '20:00:00:00')
+        timeobj = Timecode('23.98', '20:00:00:00')
         self.assertEqual('20:00:00:00', timeobj.__repr__())
 
-        timeobj = PyTimeCode('29.97', '00:09:00:00')
+        timeobj = Timecode('29.97', '00:09:00:00')
         self.assertEqual('00:08:59:28', timeobj.__repr__())
 
-        timeobj = PyTimeCode('30', '00:10:00:00')
+        timeobj = Timecode('30', '00:10:00:00')
         self.assertEqual('00:10:00:00', timeobj.__repr__())
 
-        timeobj = PyTimeCode('60', '00:00:09:00')
+        timeobj = Timecode('60', '00:00:09:00')
         self.assertEqual('00:00:09:00', timeobj.__repr__())
 
-        timeobj = PyTimeCode('59.94', '00:00:20:00')
+        timeobj = Timecode('59.94', '00:00:20:00')
         self.assertEqual('00:00:20:00', timeobj.__repr__())
 
-        timeobj = PyTimeCode('ms', '00:00:00:900')
+        timeobj = Timecode('ms', '00:00:00:900')
         self.assertEqual('00:00:00:900', timeobj.__repr__())
 
-        timeobj = PyTimeCode('24', frames=49)
+        timeobj = Timecode('24', frames=49)
         self.assertEqual('00:00:02:00', timeobj.__repr__())
 
     def test_timecode_init(self):
         """testing several timecode initialization
         """
-        tc = PyTimeCode('29.97')
+        tc = Timecode('29.97')
         self.assertEqual('00:00:00:00', tc.__str__())
         self.assertEqual(1, tc.frames)
 
-        tc = PyTimeCode('29.97', '00:00:00:01')
+        tc = Timecode('29.97', '00:00:00:01')
         self.assertEqual(2, tc.frames)
 
-        tc = PyTimeCode('29.97', '03:36:09:23')
+        tc = Timecode('29.97', '03:36:09:23')
         self.assertEqual(388704, tc.frames)
 
-        tc = PyTimeCode('29.97', '03:36:09:23')
+        tc = Timecode('29.97', '03:36:09:23')
         self.assertEqual(388704, tc.frames)
 
-        tc = PyTimeCode('30', '03:36:09:23')
+        tc = Timecode('30', '03:36:09:23')
         self.assertEqual(389094, tc.frames)
 
-        tc = PyTimeCode('25', '03:36:09:23')
+        tc = Timecode('25', '03:36:09:23')
         self.assertEqual(324249, tc.frames)
 
-        tc = PyTimeCode('59.94', '03:36:09:23')
+        tc = Timecode('59.94', '03:36:09:23')
         self.assertEqual(777384, tc.frames)
 
-        tc = PyTimeCode('60', '03:36:09:23')
+        tc = Timecode('60', '03:36:09:23')
         self.assertEqual(778164, tc.frames)
 
-        tc = PyTimeCode('59.94', '03:36:09:23')
+        tc = Timecode('59.94', '03:36:09:23')
         self.assertEqual(777384, tc.frames)
 
-        tc = PyTimeCode('23.98', '03:36:09:23')
+        tc = Timecode('23.98', '03:36:09:23')
         self.assertEqual(311280, tc.frames)
 
-        tc = PyTimeCode('24', '03:36:09:23')
+        tc = Timecode('24', '03:36:09:23')
         self.assertEqual(311280, tc.frames)
 
-        tc = PyTimeCode('ms', '03:36:09:230')
+        tc = Timecode('ms', '03:36:09:230')
         self.assertEqual(3, tc.hrs)
         self.assertEqual(36, tc.mins)
         self.assertEqual(9, tc.secs)
         self.assertEqual(230, tc.frs)
 
-        tc = PyTimeCode('24', frames=12000)
+        tc = Timecode('24', frames=12000)
         self.assertEqual('00:08:19:23', tc.__str__())
 
-        tc = PyTimeCode('29.97', frames=2589408)
+        tc = Timecode('29.97', frames=2589408)
         self.assertEqual('23:59:59:29', tc.__str__())
 
-        tc = PyTimeCode('29.97', frames=2589409)
+        tc = Timecode('29.97', frames=2589409)
         self.assertEqual('00:00:00:00', tc.__str__())
 
-        tc = PyTimeCode('59.94', frames=5178816)
+        tc = Timecode('59.94', frames=5178816)
         self.assertEqual('23:59:59:59', tc.__str__())
 
-        tc = PyTimeCode('59.94', frames=5178817)
+        tc = Timecode('59.94', frames=5178817)
         self.assertEqual('00:00:00:00', tc.__str__())
 
     def test_frame_to_tc(self):
-        tc = PyTimeCode('29.97', '00:00:00:01')
+        tc = Timecode('29.97', '00:00:00:01')
         self.assertEqual(0, tc.hrs)
         self.assertEqual(0, tc.mins)
         self.assertEqual(0, tc.secs)
         self.assertEqual(1, tc.frs)
         self.assertEqual('00:00:00:01', tc.__str__())
 
-        tc = PyTimeCode('29.97', '03:36:09:23')
+        tc = Timecode('29.97', '03:36:09:23')
         self.assertEqual(3, tc.hrs)
         self.assertEqual(36, tc.mins)
         self.assertEqual(9, tc.secs)
         self.assertEqual(23, tc.frs)
 
-        tc = PyTimeCode('29.97', '03:36:09:23')
+        tc = Timecode('29.97', '03:36:09:23')
         self.assertEqual(3, tc.hrs)
         self.assertEqual(36, tc.mins)
         self.assertEqual(9, tc.secs)
         self.assertEqual(23, tc.frs)
 
-        tc = PyTimeCode('30', '03:36:09:23')
+        tc = Timecode('30', '03:36:09:23')
         self.assertEqual(3, tc.hrs)
         self.assertEqual(36, tc.mins)
         self.assertEqual(9, tc.secs)
         self.assertEqual(23, tc.frs)
 
-        tc = PyTimeCode('25', '03:36:09:23')
+        tc = Timecode('25', '03:36:09:23')
         self.assertEqual(3, tc.hrs)
         self.assertEqual(36, tc.mins)
         self.assertEqual(9, tc.secs)
         self.assertEqual(23, tc.frs)
 
-        tc = PyTimeCode('59.94', '03:36:09:23')
+        tc = Timecode('59.94', '03:36:09:23')
         self.assertEqual(3, tc.hrs)
         self.assertEqual(36, tc.mins)
         self.assertEqual(9, tc.secs)
         self.assertEqual(23, tc.frs)
 
-        tc = PyTimeCode('60', '03:36:09:23')
+        tc = Timecode('60', '03:36:09:23')
         self.assertEqual(3, tc.hrs)
         self.assertEqual(36, tc.mins)
         self.assertEqual(9, tc.secs)
         self.assertEqual(23, tc.frs)
 
-        tc = PyTimeCode('59.94', '03:36:09:23')
+        tc = Timecode('59.94', '03:36:09:23')
         self.assertEqual(3, tc.hrs)
         self.assertEqual(36, tc.mins)
         self.assertEqual(9, tc.secs)
         self.assertEqual(23, tc.frs)
 
-        tc = PyTimeCode('23.98', '03:36:09:23')
+        tc = Timecode('23.98', '03:36:09:23')
         self.assertEqual(3, tc.hrs)
         self.assertEqual(36, tc.mins)
         self.assertEqual(9, tc.secs)
         self.assertEqual(23, tc.frs)
 
-        tc = PyTimeCode('24', '03:36:09:23')
+        tc = Timecode('24', '03:36:09:23')
         self.assertEqual(3, tc.hrs)
         self.assertEqual(36, tc.mins)
         self.assertEqual(9, tc.secs)
         self.assertEqual(23, tc.frs)
 
-        tc = PyTimeCode('ms', '03:36:09:230')
+        tc = Timecode('ms', '03:36:09:230')
         self.assertEqual(3, tc.hrs)
         self.assertEqual(36, tc.mins)
         self.assertEqual(9, tc.secs)
         self.assertEqual(230, tc.frs)
 
-        tc = PyTimeCode('24', frames=12000)
+        tc = Timecode('24', frames=12000)
         self.assertEqual('00:08:19:23', tc.__str__())
         self.assertEqual(0, tc.hrs)
         self.assertEqual(8, tc.mins)
@@ -206,49 +206,49 @@ class PyTimeCodeTester(unittest.TestCase):
     def test_tc_to_frame_test_in_2997(self):
         """testing if timecode to frame conversion is ok in 2997
         """
-        tc = PyTimeCode('29.97', '00:00:00:00')
+        tc = Timecode('29.97', '00:00:00:00')
         self.assertEqual(tc.frames, 1)
 
-        tc = PyTimeCode('29.97', '00:00:00:21')
+        tc = Timecode('29.97', '00:00:00:21')
         self.assertEqual(tc.frames, 22)
 
-        tc = PyTimeCode('29.97', '00:00:00:29')
+        tc = Timecode('29.97', '00:00:00:29')
         self.assertEqual(tc.frames, 30)
 
-        tc = PyTimeCode('29.97', '00:00:00:60')
+        tc = Timecode('29.97', '00:00:00:60')
         self.assertEqual(tc.frames, 61)
 
-        tc = PyTimeCode('29.97', '00:00:01:00')
+        tc = Timecode('29.97', '00:00:01:00')
         self.assertEqual(tc.frames, 31)
 
-        tc = PyTimeCode('29.97', '00:00:10:00')
+        tc = Timecode('29.97', '00:00:10:00')
         self.assertEqual(tc.frames, 301)
 
         # test with non existing timecodes
-        tc = PyTimeCode('29.97', '00:01:00:00')
+        tc = Timecode('29.97', '00:01:00:00')
         self.assertEqual(1799, tc.frames)
         self.assertEqual('00:00:59:28', tc.__str__())
 
         # test the limit
-        tc = PyTimeCode('29.97', '23:59:59:29')
+        tc = Timecode('29.97', '23:59:59:29')
         self.assertEqual(2589408, tc.frames)
 
     def test_drop_frame(self):
-        tc = PyTimeCode('29.97', '13:36:59:29')
+        tc = Timecode('29.97', '13:36:59:29')
         timecode = tc.next()
         self.assertEqual("13:37:00:02", timecode.__str__())
 
-        tc = PyTimeCode('59.94', '13:36:59:59')
+        tc = Timecode('59.94', '13:36:59:59')
         self.assertEqual("13:36:59:59", tc.__str__())
 
         timecode = tc.next()
         self.assertEqual("13:37:00:04", timecode.__str__())
 
-        tc = PyTimeCode('59.94', '13:39:59:59')
+        tc = Timecode('59.94', '13:39:59:59')
         timecode = tc.next()
         self.assertEqual("13:40:00:00", timecode.__str__())
 
-        tc = PyTimeCode('29.97', '13:39:59:29')
+        tc = Timecode('29.97', '13:39:59:29')
         timecode = tc.next()
         self.assertEqual("13:40:00:00", timecode.__str__())
 
@@ -256,24 +256,24 @@ class PyTimeCodeTester(unittest.TestCase):
         """testing if setting the frame rate to 29.97 forces the dropframe to
         True
         """
-        tc = PyTimeCode('29.97')
+        tc = Timecode('29.97')
         self.assertTrue(tc.drop_frame)
 
-        tc = PyTimeCode('29.97')
+        tc = Timecode('29.97')
         self.assertTrue(tc.drop_frame)
 
     def test_setting_frame_rate_to_5994_forces_drop_frame(self):
         """testing if setting the frame rate to 59.94 forces the dropframe to
         True
         """
-        tc = PyTimeCode('59.94')
+        tc = Timecode('59.94')
         self.assertTrue(tc.drop_frame)
 
-        tc = PyTimeCode('59.94')
+        tc = Timecode('59.94')
         self.assertTrue(tc.drop_frame)
 
     def test_iteration(self):
-        tc = PyTimeCode('29.97', '03:36:09:23')
+        tc = Timecode('29.97', '03:36:09:23')
         self.assertEqual("03:36:09:23", tc)
         for x in range(60):
             t = tc.next()
@@ -281,70 +281,70 @@ class PyTimeCodeTester(unittest.TestCase):
         self.assertEqual("03:36:11:23", t)
         self.assertEqual(388764, tc.frames)
 
-        tc = PyTimeCode('29.97', '03:36:09:23')
+        tc = Timecode('29.97', '03:36:09:23')
         for x in range(60):
             t = tc.next()
             self.assertTrue(t)
         self.assertEqual("03:36:11:23", t)
         self.assertEqual(388764, tc.frames)
 
-        tc = PyTimeCode('30', '03:36:09:23')
+        tc = Timecode('30', '03:36:09:23')
         for x in range(60):
             t = tc.next()
             self.assertTrue(t)
         self.assertEqual("03:36:11:23", t)
         self.assertEqual(389154, tc.frames)
 
-        tc = PyTimeCode('25', '03:36:09:23')
+        tc = Timecode('25', '03:36:09:23')
         for x in range(60):
             t = tc.next()
             self.assertTrue(t)
         self.assertEqual("03:36:12:08", t)
         self.assertEqual(324309, tc.frames)
 
-        tc = PyTimeCode('59.94', '03:36:09:23')
+        tc = Timecode('59.94', '03:36:09:23')
         for x in range(60):
             t = tc.next()
             self.assertTrue(t)
         self.assertEqual("03:36:10:23", t)
         self.assertEqual(777444, tc.frames)
 
-        tc = PyTimeCode('60', '03:36:09:23')
+        tc = Timecode('60', '03:36:09:23')
         for x in range(60):
             t = tc.next()
             self.assertTrue(t)
         self.assertEqual("03:36:10:23", t)
         self.assertEqual(778224, tc.frames)
 
-        tc = PyTimeCode('59.94', '03:36:09:23')
+        tc = Timecode('59.94', '03:36:09:23')
         for x in range(60):
             t = tc.next()
             self.assertTrue(t)
         self.assertEqual("03:36:10:23", t)
         self.assertEqual(777444, tc.frames)
 
-        tc = PyTimeCode('23.98', '03:36:09:23')
+        tc = Timecode('23.98', '03:36:09:23')
         for x in range(60):
             t = tc.next()
             self.assertTrue(t)
         self.assertEqual("03:36:12:11", t)
         self.assertEqual(311340, tc.frames)
 
-        tc = PyTimeCode('24', '03:36:09:23')
+        tc = Timecode('24', '03:36:09:23')
         for x in range(60):
             t = tc.next()
             self.assertTrue(t)
         self.assertEqual("03:36:12:11", t)
         self.assertEqual(311340, tc.frames)
 
-        tc = PyTimeCode('ms', '03:36:09:230')
+        tc = Timecode('ms', '03:36:09:230')
         for x in range(60):
             t = tc.next()
             self.assertIsNotNone(t)
         self.assertEqual('03:36:09:290', t)
         self.assertEqual(12969291, tc.frames)
 
-        tc = PyTimeCode('24', frames=12000)
+        tc = Timecode('24', frames=12000)
         for x in range(60):
             t = tc.next()
             self.assertTrue(t)
@@ -352,8 +352,8 @@ class PyTimeCodeTester(unittest.TestCase):
         self.assertEqual(12060, tc.frames)
 
     def test_op_overloads_add(self):
-        tc = PyTimeCode('29.97', '03:36:09:23')
-        tc2 = PyTimeCode('29.97', '00:00:29:23')
+        tc = Timecode('29.97', '03:36:09:23')
+        tc2 = Timecode('29.97', '00:00:29:23')
         d = tc + tc2
         f = tc + 894
         self.assertEqual("03:36:39:17", d.__str__())
@@ -361,8 +361,8 @@ class PyTimeCodeTester(unittest.TestCase):
         self.assertEqual("03:36:39:17", f.__str__())
         self.assertEqual(389598, f.frames)
 
-        tc = PyTimeCode('29.97', '03:36:09:23')
-        tc2 = PyTimeCode('29.97', '00:00:29:23')
+        tc = Timecode('29.97', '03:36:09:23')
+        tc2 = Timecode('29.97', '00:00:29:23')
         d = tc + tc2
         f = tc + 894
         self.assertEqual("03:36:39:17", d.__str__())
@@ -370,8 +370,8 @@ class PyTimeCodeTester(unittest.TestCase):
         self.assertEqual("03:36:39:17", f.__str__())
         self.assertEqual(389598, f.frames)
 
-        tc = PyTimeCode('30', '03:36:09:23')
-        tc2 = PyTimeCode('30', '00:00:29:23')
+        tc = Timecode('30', '03:36:09:23')
+        tc2 = Timecode('30', '00:00:29:23')
         d = tc + tc2
         f = tc + 894
         self.assertEqual("03:36:39:17", d.__str__())
@@ -379,8 +379,8 @@ class PyTimeCodeTester(unittest.TestCase):
         self.assertEqual("03:36:39:17", f.__str__())
         self.assertEqual(389988, f.frames)
 
-        tc = PyTimeCode('25', '03:36:09:23')
-        tc2 = PyTimeCode('25', '00:00:29:23')
+        tc = Timecode('25', '03:36:09:23')
+        tc2 = Timecode('25', '00:00:29:23')
         self.assertEqual(749, tc2.frames)
         d = tc + tc2
         f = tc + 749
@@ -389,8 +389,8 @@ class PyTimeCodeTester(unittest.TestCase):
         self.assertEqual("03:36:39:22", f.__str__())
         self.assertEqual(324998, f.frames)
 
-        tc = PyTimeCode('59.94', '03:36:09:23')
-        tc2 = PyTimeCode('59.94', '00:00:29:23')
+        tc = Timecode('59.94', '03:36:09:23')
+        tc2 = Timecode('59.94', '00:00:29:23')
         self.assertEqual(1764, tc2.frames)
         d = tc + tc2
         f = tc + 1764
@@ -399,8 +399,8 @@ class PyTimeCodeTester(unittest.TestCase):
         self.assertEqual("03:36:38:47", f.__str__())
         self.assertEqual(779148, f.frames)
 
-        tc = PyTimeCode('60', '03:36:09:23')
-        tc2 = PyTimeCode('60', '00:00:29:23')
+        tc = Timecode('60', '03:36:09:23')
+        tc2 = Timecode('60', '00:00:29:23')
         self.assertEqual(1764, tc2.frames)
         d = tc + tc2
         f = tc + 1764
@@ -409,8 +409,8 @@ class PyTimeCodeTester(unittest.TestCase):
         self.assertEqual("03:36:38:47", f.__str__())
         self.assertEqual(779928, f.frames)
 
-        tc = PyTimeCode('59.94', '03:36:09:23')
-        tc2 = PyTimeCode('59.94', '00:00:29:23')
+        tc = Timecode('59.94', '03:36:09:23')
+        tc2 = Timecode('59.94', '00:00:29:23')
         self.assertEqual(1764, tc2.frames)
         d = tc + tc2
         f = tc + 1764
@@ -419,8 +419,8 @@ class PyTimeCodeTester(unittest.TestCase):
         self.assertEqual("03:36:38:47", f.__str__())
         self.assertEqual(779148, f.frames)
 
-        tc = PyTimeCode('23.98', '03:36:09:23')
-        tc2 = PyTimeCode('23.98', '00:00:29:23')
+        tc = Timecode('23.98', '03:36:09:23')
+        tc2 = Timecode('23.98', '00:00:29:23')
         self.assertEqual(720, tc2.frames)
         d = tc + tc2
         f = tc + 720
@@ -429,8 +429,8 @@ class PyTimeCodeTester(unittest.TestCase):
         self.assertEqual("03:36:39:23", f.__str__())
         self.assertEqual(312000, f.frames)
 
-        tc = PyTimeCode('23.98', '03:36:09:23')
-        tc2 = PyTimeCode('23.98', '00:00:29:23')
+        tc = Timecode('23.98', '03:36:09:23')
+        tc2 = Timecode('23.98', '00:00:29:23')
         self.assertEqual(720, tc2)
         d = tc + tc2
         f = tc + 720
@@ -439,8 +439,8 @@ class PyTimeCodeTester(unittest.TestCase):
         self.assertEqual("03:36:39:23", f.__str__())
         self.assertEqual(312000, f.frames)
 
-        tc = PyTimeCode('ms', '03:36:09:230')
-        tc2 = PyTimeCode('ms', '01:06:09:230')
+        tc = Timecode('ms', '03:36:09:230')
+        tc2 = Timecode('ms', '01:06:09:230')
         self.assertEqual(3969231, tc2.frames)
         d = tc + tc2
         f = tc + 720
@@ -449,8 +449,8 @@ class PyTimeCodeTester(unittest.TestCase):
         self.assertEqual("03:36:09:950", f.__str__())
         self.assertEqual(12969951, f.frames)
 
-        tc = PyTimeCode('24', frames=12000)
-        tc2 = PyTimeCode('24', frames=485)
+        tc = Timecode('24', frames=12000)
+        tc2 = Timecode('24', frames=485)
         self.assertEqual(485, tc2.frames)
         d = tc + tc2
         f = tc + 719
@@ -463,48 +463,48 @@ class PyTimeCodeTester(unittest.TestCase):
         """testing if the resultant object will have the left sides frame rate
         when two timecodes with different frame rates are added together
         """
-        tc1 = PyTimeCode('29.97', '00:00:00:00')
-        tc2 = PyTimeCode('24', '00:00:00:10')
+        tc1 = Timecode('29.97', '00:00:00:00')
+        tc2 = Timecode('24', '00:00:00:10')
         tc3 = tc1 + tc2
         self.assertEqual('29.97', tc3.framerate)
         self.assertEqual(12, tc3.frames)
         self.assertEqual('00:00:00:11', tc3)
 
     def test_frame_number_attribute_value_is_correctly_calculated(self):
-        """testing if the PyTimeCode.frame_number attribute is correctly
+        """testing if the Timecode.frame_number attribute is correctly
         calculated
         """
-        tc1 = PyTimeCode('24', '00:00:00:00')
+        tc1 = Timecode('24', '00:00:00:00')
         self.assertEqual(1, tc1.frames)
         self.assertEqual(0, tc1.frame_number)
 
-        tc2 = PyTimeCode('24', '00:00:01:00')
+        tc2 = Timecode('24', '00:00:01:00')
         self.assertEqual(25, tc2.frames)
         self.assertEqual(24, tc2.frame_number)
 
-        tc3 = PyTimeCode('29.97', '00:01:00:00')
+        tc3 = Timecode('29.97', '00:01:00:00')
         self.assertEqual(1799, tc3.frames)
         self.assertEqual(1798, tc3.frame_number)
 
-        tc4 = PyTimeCode('30', '00:01:00:00')
+        tc4 = Timecode('30', '00:01:00:00')
         self.assertEqual(1801, tc4.frames)
         self.assertEqual(1800, tc4.frame_number)
 
-        tc5 = PyTimeCode('50', '00:01:00:00')
+        tc5 = Timecode('50', '00:01:00:00')
         self.assertEqual(3001, tc5.frames)
         self.assertEqual(3000, tc5.frame_number)
 
-        tc6 = PyTimeCode('59.94', '00:01:00:00')
+        tc6 = Timecode('59.94', '00:01:00:00')
         self.assertEqual(3597, tc6.frames)
         self.assertEqual(3596, tc6.frame_number)
 
-        tc7 = PyTimeCode('60', '00:01:00:00')
+        tc7 = Timecode('60', '00:01:00:00')
         self.assertEqual(3601, tc7.frames)
         self.assertEqual(3600, tc7.frame_number)
 
     def test_op_overloads_subtract(self):
-        tc = PyTimeCode('29.97', '03:36:09:23')
-        tc2 = PyTimeCode('29.97', '00:00:29:23')
+        tc = Timecode('29.97', '03:36:09:23')
+        tc2 = Timecode('29.97', '00:00:29:23')
         self.assertEqual(894, tc2.frames)
         d = tc - tc2
         f = tc - 894
@@ -513,8 +513,8 @@ class PyTimeCodeTester(unittest.TestCase):
         self.assertEqual("03:35:39:27", f.__str__())
         self.assertEqual(387810, f.frames)
 
-        tc = PyTimeCode('29.97', '03:36:09:23')
-        tc2 = PyTimeCode('29.97', '00:00:29:23')
+        tc = Timecode('29.97', '03:36:09:23')
+        tc2 = Timecode('29.97', '00:00:29:23')
         d = tc - tc2
         f = tc - 894
         self.assertEqual("03:35:39:27", d.__str__())
@@ -522,8 +522,8 @@ class PyTimeCodeTester(unittest.TestCase):
         self.assertEqual("03:35:39:27", f.__str__())
         self.assertEqual(387810, f.frames)
 
-        tc = PyTimeCode('30', '03:36:09:23')
-        tc2 = PyTimeCode('30', '00:00:29:23')
+        tc = Timecode('30', '03:36:09:23')
+        tc2 = Timecode('30', '00:00:29:23')
         d = tc - tc2
         f = tc - 894
         self.assertEqual("03:35:39:29", d.__str__())
@@ -531,8 +531,8 @@ class PyTimeCodeTester(unittest.TestCase):
         self.assertEqual("03:35:39:29", f.__str__())
         self.assertEqual(388200, f.frames)
 
-        tc = PyTimeCode('25', '03:36:09:23')
-        tc2 = PyTimeCode('25', '00:00:29:23')
+        tc = Timecode('25', '03:36:09:23')
+        tc2 = Timecode('25', '00:00:29:23')
         self.assertEqual(749, tc2.frames)
         d = tc - tc2
         f = tc - 749
@@ -541,8 +541,8 @@ class PyTimeCodeTester(unittest.TestCase):
         self.assertEqual("03:35:39:24", f.__str__())
         self.assertEqual(323500, f.frames)
 
-        tc = PyTimeCode('59.94', '03:36:09:23')
-        tc2 = PyTimeCode('59.94', '00:00:29:23')
+        tc = Timecode('59.94', '03:36:09:23')
+        tc2 = Timecode('59.94', '00:00:29:23')
         self.assertEqual(1764, tc2.frames)
         d = tc - tc2
         f = tc - 1764
@@ -551,8 +551,8 @@ class PyTimeCodeTester(unittest.TestCase):
         self.assertEqual("03:35:39:55", f.__str__())
         self.assertEqual(775620, f.frames)
 
-        tc = PyTimeCode('60', '03:36:09:23')
-        tc2 = PyTimeCode('60', '00:00:29:23')
+        tc = Timecode('60', '03:36:09:23')
+        tc2 = Timecode('60', '00:00:29:23')
         self.assertEqual(1764, tc2.frames)
         d = tc - tc2
         f = tc - 1764
@@ -561,8 +561,8 @@ class PyTimeCodeTester(unittest.TestCase):
         self.assertEqual("03:35:39:59", f.__str__())
         self.assertEqual(776400, f.frames)
 
-        tc = PyTimeCode('59.94', '03:36:09:23')
-        tc2 = PyTimeCode('59.94', '00:00:29:23')
+        tc = Timecode('59.94', '03:36:09:23')
+        tc2 = Timecode('59.94', '00:00:29:23')
         d = tc - tc2
         f = tc - 1764
         self.assertEqual("03:35:39:55", d.__str__())
@@ -570,8 +570,8 @@ class PyTimeCodeTester(unittest.TestCase):
         self.assertEqual("03:35:39:55", f.__str__())
         self.assertEqual(775620, f.frames)
 
-        tc = PyTimeCode('23.98', '03:36:09:23')
-        tc2 = PyTimeCode('23.98', '00:00:29:23')
+        tc = Timecode('23.98', '03:36:09:23')
+        tc2 = Timecode('23.98', '00:00:29:23')
         self.assertEqual(720, tc2.frames)
         d = tc - tc2
         f = tc - 720
@@ -580,8 +580,8 @@ class PyTimeCodeTester(unittest.TestCase):
         self.assertEqual("03:35:39:23", f.__str__())
         self.assertEqual(310560, f.frames)
 
-        tc = PyTimeCode('23.98', '03:36:09:23')
-        tc2 = PyTimeCode('23.98', '00:00:29:23')
+        tc = Timecode('23.98', '03:36:09:23')
+        tc2 = Timecode('23.98', '00:00:29:23')
         d = tc - tc2
         f = tc - 720
         self.assertEqual("03:35:39:23", d.__str__())
@@ -589,8 +589,8 @@ class PyTimeCodeTester(unittest.TestCase):
         self.assertEqual("03:35:39:23", f.__str__())
         self.assertEqual(310560, f.frames)
 
-        tc = PyTimeCode('ms', '03:36:09:230')
-        tc2 = PyTimeCode('ms', '01:06:09:230')
+        tc = Timecode('ms', '03:36:09:230')
+        tc2 = Timecode('ms', '01:06:09:230')
         self.assertEqual(3969231, tc2.frames)
         d = tc - tc2
         f = tc - 3969231
@@ -599,8 +599,8 @@ class PyTimeCodeTester(unittest.TestCase):
         self.assertEqual("02:29:59:999", f.__str__())
         self.assertEqual(9000000, f.frames)
 
-        tc = PyTimeCode('24', frames=12000)
-        tc2 = PyTimeCode('24', frames=485)
+        tc = Timecode('24', frames=12000)
+        tc2 = Timecode('24', frames=485)
         self.assertEqual(485, tc2.frames)
         d = tc - tc2
         f = tc - 485
@@ -610,8 +610,8 @@ class PyTimeCodeTester(unittest.TestCase):
         self.assertEqual(11515, f.frames)
 
     def test_op_overloads_mult(self):
-        tc = PyTimeCode('29.97', '00:00:09:23')
-        tc2 = PyTimeCode('29.97', '00:00:29:23')
+        tc = Timecode('29.97', '00:00:09:23')
+        tc2 = Timecode('29.97', '00:00:29:23')
         d = tc * tc2
         f = tc * 4
         self.assertEqual("02:26:09:29", d.__str__())
@@ -619,8 +619,8 @@ class PyTimeCodeTester(unittest.TestCase):
         self.assertEqual("00:00:39:05", f.__str__())
         self.assertEqual(1176, f.frames)
 
-        tc = PyTimeCode('29.97', '00:00:09:23')
-        tc2 = PyTimeCode('29.97', '00:00:29:23')
+        tc = Timecode('29.97', '00:00:09:23')
+        tc2 = Timecode('29.97', '00:00:29:23')
         d = tc * tc2
         f = tc * 4
         self.assertEqual("02:26:09:29", d.__str__())
@@ -628,8 +628,8 @@ class PyTimeCodeTester(unittest.TestCase):
         self.assertEqual("00:00:39:05", f.__str__())
         self.assertEqual(1176, f.frames)
 
-        tc = PyTimeCode('30', '03:36:09:23')
-        tc2 = PyTimeCode('30', '00:00:29:23')
+        tc = Timecode('30', '03:36:09:23')
+        tc2 = Timecode('30', '00:00:29:23')
         d = tc * tc2
         f = tc * 894
         self.assertEqual("04:50:01:05", d.__str__())
@@ -637,8 +637,8 @@ class PyTimeCodeTester(unittest.TestCase):
         self.assertEqual("04:50:01:05", f.__str__())
         self.assertEqual(347850036, f.frames)
 
-        tc = PyTimeCode('25', '03:36:09:23')
-        tc2 = PyTimeCode('25', '00:00:29:23')
+        tc = Timecode('25', '03:36:09:23')
+        tc2 = Timecode('25', '00:00:29:23')
         self.assertEqual(749, tc2.frames)
         d = tc * tc2
         f = tc * 749
@@ -647,8 +647,8 @@ class PyTimeCodeTester(unittest.TestCase):
         self.assertEqual("10:28:20:00", f.__str__())
         self.assertEqual(242862501, f.frames)
 
-        tc = PyTimeCode('59.94', '03:36:09:23')
-        tc2 = PyTimeCode('59.94', '00:00:29:23')
+        tc = Timecode('59.94', '03:36:09:23')
+        tc2 = Timecode('59.94', '00:00:29:23')
         self.assertEqual(1764, tc2.frames)
         d = tc * tc2
         f = tc * 1764
@@ -657,8 +657,8 @@ class PyTimeCodeTester(unittest.TestCase):
         self.assertEqual("18:59:27:35", f.__str__())
         self.assertEqual(1371305376, f.frames)
 
-        tc = PyTimeCode('60', '03:36:09:23')
-        tc2 = PyTimeCode('60', '00:00:29:23')
+        tc = Timecode('60', '03:36:09:23')
+        tc2 = Timecode('60', '00:00:29:23')
         self.assertEqual(1764, tc2.frames)
         d = tc * tc2
         f = tc * 1764
@@ -667,8 +667,8 @@ class PyTimeCodeTester(unittest.TestCase):
         self.assertEqual("19:00:21:35", f.__str__())
         self.assertEqual(1372681296, f.frames)
 
-        tc = PyTimeCode('59.94', '03:36:09:23')
-        tc2 = PyTimeCode('59.94', '00:00:29:23')
+        tc = Timecode('59.94', '03:36:09:23')
+        tc2 = Timecode('59.94', '00:00:29:23')
         d = tc * tc2
         f = tc * 1764
         self.assertEqual("18:59:27:35", d.__str__())
@@ -676,8 +676,8 @@ class PyTimeCodeTester(unittest.TestCase):
         self.assertEqual("18:59:27:35", f.__str__())
         self.assertEqual(1371305376, f.frames)
 
-        tc = PyTimeCode('23.98', '03:36:09:23')
-        tc2 = PyTimeCode('23.98', '00:00:29:23')
+        tc = Timecode('23.98', '03:36:09:23')
+        tc2 = Timecode('23.98', '00:00:29:23')
         self.assertEqual(720, tc2.frames)
         d = tc * tc2
         f = tc * 720
@@ -686,8 +686,8 @@ class PyTimeCodeTester(unittest.TestCase):
         self.assertEqual("04:09:35:23", f.__str__())
         self.assertEqual(224121600, f.frames)
 
-        tc = PyTimeCode('ms', '03:36:09:230')
-        tc2 = PyTimeCode('ms', '01:06:09:230')
+        tc = Timecode('ms', '03:36:09:230')
+        tc2 = Timecode('ms', '01:06:09:230')
         self.assertEqual(3969231, tc2.frames)
         d = tc * tc2
         f = tc * 3969231
@@ -696,8 +696,8 @@ class PyTimeCodeTester(unittest.TestCase):
         self.assertEqual("17:22:11:360", f.__str__())
         self.assertEqual(51477873731361, f.frames)
 
-        tc = PyTimeCode('24', frames=12000)
-        tc2 = PyTimeCode('24', frames=485)
+        tc = Timecode('24', frames=12000)
+        tc2 = Timecode('24', frames=485)
         self.assertEqual(485, tc2.frames)
         d = tc * tc2
         f = tc * 485
@@ -710,8 +710,8 @@ class PyTimeCodeTester(unittest.TestCase):
         """testing if the timecode will loop back to 00:00:00:00 after 24 hours
         in 24 fps
         """
-        tc = PyTimeCode('24', '00:00:00:21')
-        tc2 = PyTimeCode('24', '23:59:59:23')
+        tc = Timecode('24', '00:00:00:21')
+        tc2 = Timecode('24', '23:59:59:23')
         self.assertEqual(
             '00:00:00:21',
             (tc + tc2).__str__()
@@ -725,11 +725,11 @@ class PyTimeCodeTester(unittest.TestCase):
         """testing if the timecode will loop back to 00:00:00:00 after 24 hours
         in 29.97 fps
         """
-        tc = PyTimeCode('29.97', '00:00:00:21')
+        tc = Timecode('29.97', '00:00:00:21')
         self.assertTrue(tc.drop_frame)
         self.assertEqual(22, tc.frames)
 
-        tc2 = PyTimeCode('29.97', '23:59:59:29')
+        tc2 = Timecode('29.97', '23:59:59:29')
         self.assertTrue(tc2.drop_frame)
         self.assertEqual(2589408, tc2.frames)
 
@@ -766,133 +766,133 @@ class PyTimeCodeTester(unittest.TestCase):
         """testing if the timecode will loop back to 00:00:00:00 after 24 hours
         in 29.97 fps
         """
-        tc0 = PyTimeCode('59.94', '23:59:59:29')
+        tc0 = Timecode('59.94', '23:59:59:29')
         self.assertEqual(5178786, tc0.frames)
-        tc0 = PyTimeCode('29.97', '23:59:59:29')
+        tc0 = Timecode('29.97', '23:59:59:29')
         self.assertEqual(2589408, tc0.frames)
 
-        tc1 = PyTimeCode('29.97', frames=2589408)
+        tc1 = Timecode('29.97', frames=2589408)
         self.assertEqual('23:59:59:29', tc1.__str__())
 
-        tc2 = PyTimeCode('29.97', '23:59:59:29')
+        tc2 = Timecode('29.97', '23:59:59:29')
         tc3 = tc2 + 1
         self.assertEqual('00:00:00:00', tc3.__str__())
 
-        tc2 = PyTimeCode('29.97', '23:59:59:29')
+        tc2 = Timecode('29.97', '23:59:59:29')
         tc3 = tc2 + 21
         self.assertEqual('00:00:00:20', tc3.__str__())
 
-        tc = PyTimeCode('29.97', '00:00:00:21')
-        tc2 = PyTimeCode('29.97', '23:59:59:29')
+        tc = Timecode('29.97', '00:00:00:21')
+        tc2 = Timecode('29.97', '23:59:59:29')
         tc3 = (tc + tc2)
         self.assertEqual('00:00:00:21', tc3.__str__())
 
-        tc = PyTimeCode('29.97', '04:20:13:21')
-        tca = PyTimeCode('29.97', frames=467944)
+        tc = Timecode('29.97', '04:20:13:21')
+        tca = Timecode('29.97', frames=467944)
         self.assertEqual(467944, tca.frames)
         self.assertEqual(467944, tc.frames)
         self.assertEqual('04:20:13:21', tca.__str__())
         self.assertEqual('04:20:13:21', tc.__str__())
 
-        tc2 = PyTimeCode('29.97', '23:59:59:29')
+        tc2 = Timecode('29.97', '23:59:59:29')
         self.assertEqual(2589408, tc2.frames)
         self.assertEqual('23:59:59:29', tc2.__str__())
-        tc2a = PyTimeCode('29.97', frames=2589408)
+        tc2a = Timecode('29.97', frames=2589408)
         self.assertEqual(2589408, tc2a.frames)
         self.assertEqual('23:59:59:29', tc2a.__str__())
 
         tc3 = (tc + tc2)
         self.assertEqual('04:20:13:21', tc3.__str__())
 
-        tc = PyTimeCode('59.94', '04:20:13:21')
+        tc = Timecode('59.94', '04:20:13:21')
         self.assertEqual('04:20:13:21', tc.__str__())
 
-        tca = PyTimeCode('59.94', frames=935866)
+        tca = Timecode('59.94', frames=935866)
         self.assertEqual('04:20:13:21', tc.__str__())
 
-        tc2 = PyTimeCode('59.94', '23:59:59:59')
+        tc2 = Timecode('59.94', '23:59:59:59')
         tc3 = (tc + tc2)
         self.assertEqual('04:20:13:21', tc3.__str__())
 
     def test_exceptions(self):
         """test exceptions
         """
-        # with self.assertRaises(PyTimeCodeError) as cm:
-        #     PyTimeCode('23.98', '01:20:30:303')
+        # with self.assertRaises(TimecodeError) as cm:
+        #     Timecode('23.98', '01:20:30:303')
 
         # self.assertEqual(
         #     'Timecode string parsing error. 01:20:30:303',
         #     cm.exception.__str__()
         # )
 
-        # with self.assertRaises(PyTimeCodeError) as cm:
-        #     PyTimeCode('24', '01:20:30:303')
+        # with self.assertRaises(TimecodeError) as cm:
+        #     Timecode('24', '01:20:30:303')
         # 
         # self.assertEqual(
         #     'Timecode string parsing error. 01:20:30:303',
         #     cm.exception.__str__()
         # )
         # 
-        # with self.assertRaises(PyTimeCodeError) as cm:
-        #     PyTimeCode('29.97', '01:20:30:303')
-        # 
-        # self.assertEqual(
-        #     'Timecode string parsing error. 01:20:30:303',
-        #     cm.exception.__str__()
-        # )
-
-        # with self.assertRaises(PyTimeCodeError) as cm:
-        #     PyTimeCode('30', '01:20:30:303')
+        # with self.assertRaises(TimecodeError) as cm:
+        #     Timecode('29.97', '01:20:30:303')
         # 
         # self.assertEqual(
         #     'Timecode string parsing error. 01:20:30:303',
         #     cm.exception.__str__()
         # )
 
-        # with self.assertRaises(PyTimeCodeError) as cm:
-        #     PyTimeCode('59.94', '01:20:30:303')
+        # with self.assertRaises(TimecodeError) as cm:
+        #     Timecode('30', '01:20:30:303')
+        # 
+        # self.assertEqual(
+        #     'Timecode string parsing error. 01:20:30:303',
+        #     cm.exception.__str__()
+        # )
+
+        # with self.assertRaises(TimecodeError) as cm:
+        #     Timecode('59.94', '01:20:30:303')
         # 
         # self.assertEqual(
         #     'Timecode string parsing error. 01:20:30:303',
         #     cm.exception.__str__()
         # )
         # 
-        # with self.assertRaises(PyTimeCodeError) as cm:
-        #     PyTimeCode('60', '01:20:30:303')
+        # with self.assertRaises(TimecodeError) as cm:
+        #     Timecode('60', '01:20:30:303')
         # 
         # self.assertEqual(
         #     'Timecode string parsing error. 01:20:30:303',
         #     cm.exception.__str__()
         # )
         
-        # with self.assertRaises(PyTimeCodeError) as cm:
-        #     PyTimeCode('ms', '01:20:30:3039')
+        # with self.assertRaises(TimecodeError) as cm:
+        #     Timecode('ms', '01:20:30:3039')
         # 
         # self.assertEqual(
         #     'Timecode string parsing error. 01:20:30:3039',
         #     cm.exception.__str__()
         # )
         # 
-        # with self.assertRaises(PyTimeCodeError) as cm:
-        #     PyTimeCode('60', '01:20:30:30')
+        # with self.assertRaises(TimecodeError) as cm:
+        #     Timecode('60', '01:20:30:30')
         # 
         # self.assertEqual(
         #     'Drop frame with 60fps not supported, only 29.97 & 59.94.',
         #     cm.exception.__str__(),
         # )
         # 
-        # tc = PyTimeCode('29.97', '00:00:09:23')
+        # tc = Timecode('29.97', '00:00:09:23')
         # tc2 = 'bum'
-        # with self.assertRaises(PyTimeCodeError) as cm:
+        # with self.assertRaises(TimecodeError) as cm:
         #     d = tc * tc2
         # self.assertEqual(
         #     "Type str not supported for arithmetic.",
         #     cm.exception.__str__()
         # )
         
-        # tc = PyTimeCode('30', '00:00:09:23')
+        # tc = Timecode('30', '00:00:09:23')
         # tc2 = 'bum'
-        # with self.assertRaises(PyTimeCodeError) as cm:
+        # with self.assertRaises(TimecodeError) as cm:
         #     d = tc + tc2
         # 
         # self.assertEqual(
@@ -900,9 +900,9 @@ class PyTimeCodeTester(unittest.TestCase):
         #     cm.exception.__str__()
         # )
         # 
-        # tc = PyTimeCode('24', '00:00:09:23')
+        # tc = Timecode('24', '00:00:09:23')
         # tc2 = 'bum'
-        # with self.assertRaises(PyTimeCodeError) as cm:
+        # with self.assertRaises(TimecodeError) as cm:
         #     d = tc - tc2
         # 
         # self.assertEqual(
@@ -910,9 +910,9 @@ class PyTimeCodeTester(unittest.TestCase):
         #     cm.exception.__str__()
         # )
         # 
-        # tc = PyTimeCode('ms', '00:00:09:237')
+        # tc = Timecode('ms', '00:00:09:237')
         # tc2 = 'bum'
-        # with self.assertRaises(PyTimeCodeError) as cm:
+        # with self.assertRaises(TimecodeError) as cm:
         #     d = tc / tc2
         # 
         # self.assertEqual(
